only run within embedded chat
Before this patch, the extension was running twice on /bigscreen: once on the
bigscreen page itself and again in the embedded chat `iframe`. The instance on
the bigscreen page just errors out because `destiny.chat` is `undefined`, but
the instance in the `iframe` continues working normally which is probably how
this was never noticed.