New phrases feature
Prevents users from sending messages that match a given regex

This is just a basic 'proof of concept' implementation, the actual list of bad phrases has to be added yet. The handling should also be improved - right now the message just stays in the chat input box with no additional feedback.